Issue #440 - Logging for /script/upcp and /script/cleanup. - /script/upcp does not start anymore the cleanup process if there is no new Kloxo version. If a user want to fix problems the user can run /script/cleanup manualy. - Refactored some functions from the update/cleanup process - upcp/cleanup logs to screen and to the /kloxo/logs/update file. - Removed not used/empty functions  - All above items are tested.

+++ kloxo/httpdocs/htmllib/lib/commonfslib.php
@@ -382,8 +382,12 @@ function lxfile_cp($src, $dst)
 		$dst = $dst . "/" . basename($src);
 	}
 
-	log_filesys("Copying $src $dst");
-	system("cp $src $dst", $ret);
+	if(lxfile_exists($src)) {
+		log_filesys("Copying $src $dst");
+		system("cp $src $dst", $ret);
+	} else {
+		log_filesys("ERROR! Copy $src to $dst failed!");
+	}
 	return $ret == 0;
 }
